+++ packages/infinity/src/components/util/InfinitySearch.vue
@@ -27,24 +27,6 @@
     </template>
     <v-card>
       <v-list dense>
-        <template v-if="reportResults.length">
-          <v-subheader>REPORTS</v-subheader>
-          <v-list-item
-            v-for="(result, index) in reportResults"
-            :key="`report-${index}`"
-            @click="navigateToReport(result)"
-          >
-            <v-list-item-icon>
-              <v-icon>mdi-file-chart-outline</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title>
-                {{ result.reportDescription }}
-              </v-list-item-title>
-            </v-list-item-content>
-          </v-list-item>
-        </template>
-        <v-divider v-if="reportResults.length && insightResults.length"></v-divider>
         <template v-if="insightResults.length">
           <v-subheader>INSIGHTS</v-subheader>
           <v-list-item
@@ -63,36 +45,6 @@
           </v-list-item>
         </template>
         <template v-if="noText">
-          <v-subheader>MOST VIEWED REPORTS</v-subheader>
-          <v-list-item
-            v-for="(result, index) in reports"
-            :key="`mreport-${index}`"
-            @click="navigateToReport(result)"
-          >
-            <v-list-item-icon>
-              <v-icon>mdi-file-chart-outline</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title>
-                {{ result.reportDescription }}
-              </v-list-item-title>
-            </v-list-item-content>
-          </v-list-item>
-          <v-subheader>RECENT REPORTS</v-subheader>
-          <v-list-item
-            v-for="(result, index) in reports"
-            :key="`rreport-${index}`"
-            @click="navigateToReport(result)"
-          >
-            <v-list-item-icon>
-              <v-icon>mdi-file-chart-outline</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title>
-                {{ result.reportDescription }}
-              </v-list-item-title>
-            </v-list-item-content>
-          </v-list-item>
           <v-subheader>INSIGHTS</v-subheader>
           <v-list-item
             v-for="(result, index) in insights"
@@ -125,7 +77,7 @@
 
 <script>
 import FlexSearch from 'flexsearch';
-import { mapMutations } from 'vuex';
+import { mapState, mapMutations, mapActions } from 'vuex';
 
 export default {
   name: 'InfinitySearch',
@@ -136,44 +88,7 @@ export default {
       isFocused: false,
       isSearching: false,
       insight: null,
-      report: null,
       menu: null,
-      insights: [
-        {
-          name: 'What were the top reasons for downtime in the last week?',
-          reportName: 1,
-        },
-        {
-          name: 'What were the total parts produced last week, machine wise?',
-          reportName: 2,
-        },
-        {
-          name: 'What is the summary of yesterdays production?',
-          reportName: 3,
-        },
-        {
-          name: 'What is the trend of OEE over last 7 days?',
-          reportName: 4,
-        },
-      ],
-      reports: [
-        {
-          reportName: 'productionbymachine',
-          reportDescription: 'Production By Machine',
-        },
-        {
-          reportName: 'downtimebyreason',
-          reportDescription: 'Downtime By Reasons',
-        },
-        {
-          reportName: 'oeebymonth',
-          reportDescription: 'OEE By Month',
-        },
-        {
-          reportName: 'rejectionbyplan',
-          reportDescription: 'Rejection By Plan',
-        },
-      ],
     };
   },
   beforeMount() {
@@ -183,14 +98,15 @@ export default {
         field: 'name',
       },
     });
-    this.report = new FlexSearch({
-      doc: {
-        id: 'reportName',
-        field: 'reportDescription',
-      },
-    });
     this.insight.add(this.insights);
-    this.report.add(this.reports);
+  },
+  watch: {
+    insights: {
+      deep: true,
+      handler() {
+        this.insight.add(this.insights);
+      },
+    },
   },
   mounted() {
     document.onkeydown = (event) => {
@@ -205,11 +121,11 @@ export default {
     document.onkeydown = null;
   },
   computed: {
+    ...mapState('insight', ['insights']),
     results() {
       if (this.insight === null || this.search === null) {
         return {
           insightResults: [],
-          reportResults: [],
         };
       }
       const insightResults = this.insight.search({
@@ -217,14 +133,8 @@ export default {
         suggest: true,
         limit: 3,
       });
-      const reportResults = this.report.search({
-        query: this.search,
-        suggest: true,
-        limit: 3,
-      });
       return {
         insightResults,
-        reportResults,
       };
     },
     insightResults() {
@@ -233,29 +143,24 @@ export default {
       }
       return [];
     },
-    reportResults() {
-      if (this.results && this.results.reportResults && this.results.reportResults.length) {
-        return this.results.reportResults;
-      }
-      return [];
-    },
     noText() {
       return !this.search;
     },
     noResults() {
-      return !this.insightResults.length && !this.reportResults.length;
+      return !this.insightResults.length;
     },
   },
   methods: {
-    ...mapMutations('insight', ['setQuery', 'setWindow']),
+    ...mapMutations('insight', ['setQuery', 'setWindow', 'setLoading']),
     ...mapMutations('helper', ['setInsightsDrawer']),
-    executeInsight(query) {
+    ...mapActions('insight', ['fetchInsightDetails']),
+    async executeInsight(query) {
+      this.setQuery(query);
       this.setWindow(1);
       this.setInsightsDrawer(true);
-      this.setQuery(query);
-    },
-    navigateToReport(report) {
-      this.$router.push({ name: 'reports', params: { id: report.reportName } });
+      this.setLoading(true);
+      await this.fetchInsightDetails();
+      this.setLoading(false);
     },
     onBlur() {
       this.resetSearch();